Update how artificial specs are created/loaded
In this setup of artificial requirements, the artificial .slx files contain no information
that exists in the original .slx files. As a result, the final spec that we want to falsify
is not fully contained in the artificial .slx file, but it requires use of BOTH the original
and the artificial .slx.

As an example, to create the artificial requirement ADI_req_art for falsification, we need
 - ADI.slx (for outport ADI_req)
 - ADI_art.slx (for outport ADI_art)

When both files ADI_req.stl and ADI_art.stl are generated, the final spec used for falsification
(which is called phi_ADI_req_art) is simply '(ADI_req) or (ADI_art)'.

In this way, if we perform a change in a specific original .slx (like ADI.slx), we don't need to
copy the change into its corresponding artificial .slx. This was not the case before. The downside
is that createAndLoadRequirements.m becomes a little bit more complex.