New artificial specs: "Kill the corners"
For the new versions of the artificial specs, we need the signals to be
"in the middle of the interval" instead of "in the bottom" to falsify the spec.

As an example, the artificial signals have range [0 100]. Before, to falsify the
spec we could need that the signal should be lower than 10 to falsify the spec.
Now, we instead need it to be e.g. in the interval [45 55] to falsify the spec.

This is to make it so that no corner sample falsifies the spec.